requirements.txt
django==5.2.4
gunicorn>=21.2.0
whitenoise>=6.7.0
dj-database-url>=2.2.0
psycopg2-binary>=2.9.9
djangorestframework>=3.15.2
drf-spectacular>=0.27.2
